+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.28b2" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2025-03-18"
+    AppVersionStr = "0.29b1" # see https://peps.python.org/pep-0440/
+    AppDateStr = "2025-04-03"