+++ src/lib/expressions/expressions.py
@@ -31,7 +31,6 @@ class ExpressionParser:
                 for nw in network_list:
                     if pattern is None or fnmatch.fnmatch(nw.filename, pattern):
                         nws.append(nw)
-            logging.debug(f'Expression pattern "{pattern}" matched "' + '", "'.join([n.name for n in nws]) + '"')
             if single:
                 if len(nws) != 1:
                     #raise RuntimeError(f'The pattern "{pattern}" matched {len(nws)} networks, but need exactly one')